pyproject.toml
[project]
name = "emailassitant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.116.1",
    "imapclient>=3.0.1",
    "mcp[cli]>=1.10.1",
    "pyrogram>=2.0.106",
    "python-telegram-bot>=22.2",
    "sentence-transformers>=5.0.0",
    "sqlite-vec>=0.1.6",
    "streamlit>=1.47.0",
    "telethon>=1.40.0",
    "tgcrypto>=1.2.5",
]
